OOP-java: Learned about Constructor
this keywords

object instance

class is a template/blueprint of any instances

class is a logical implication & object is a physical implication

eg: class Car is template of a Car, having engine, model name, price. Object of that class is a brand who's making physical car with the help of Blueprint, that include engine power, model name & price of the car. That is initialized by the Car manufacturer (object name that we given).